code base re-organized and all tests passed
