+++ src/client.ts
@@ -1,9 +1,9 @@
 // cwco, copyright (c) by Elson Correia
 // Distributed under an MIT license: https://github.com/beforesemicolon/cwco/blob/master/LICENSE
 
-import {WebComponent} from './web-component';
-import {Directive} from './directive';
-import {ContextProviderComponent} from './context-provider-component';
+import {WebComponent} from './core/web-component';
+import {Directive} from './core/directive';
+import {ContextProviderComponent} from './core/context-provider-component';
 import {html} from "./utils/html";
 
 // @ts-ignore
