+++ src/core/boolean-attributes.json
