+++ src/core/metadata.ts
