+++ src/core/utils/get-style-string.spec.ts
