+++ src/core/utils/get-style-string.ts
