+++ src/core/utils/set-component-properties-from-observed-attributes.spec.ts
