+++ src/core/utils/setup-component-properties-for-auto-update.spec.ts
