+++ src/core/web-component.spec.ts
@@ -1,5 +1,5 @@
 import {WebComponent} from './web-component';
-import {ShadowRootModeExtended} from "./enums/ShadowRootModeExtended.enum";
+import {ShadowRootModeExtended} from "../enums/ShadowRootModeExtended.enum";
 
 describe('WebComponent', () => {
