+++ src/directives/if.directive.ts
@@ -1,4 +1,4 @@
-import {Directive} from "../directive";
+import {Directive} from "../core/directive";
 import {CWCO} from "../cwco";
 
 export class If extends Directive {
