+++ src/enums/track-type.ts
@@ -1,0 +1,5 @@
+export enum TrackType {
+	attribute,
+	property,
+	directive
+}
