+++ src/index.ts
@@ -1,8 +1,8 @@
 // cwco, copyright (c) by Elson Correia
 // Distributed under an MIT license: https://github.com/beforesemicolon/cwco/blob/master/LICENSE
-export {WebComponent} from './web-component';
-export {ContextProviderComponent} from './context-provider-component';
-export {Directive} from './directive';
+export {WebComponent} from './core/web-component';
+export {ContextProviderComponent} from './core/context-provider-component';
+export {Directive} from './core/directive';
 export {CWCO} from './cwco';
 export {ShadowRootModeExtended} from './enums/ShadowRootModeExtended.enum';
 export {html} from './utils/html';
