+++ src/parser/self-closing-tags.json
