+++ src/tracker/utils/define-node-context-metadata.ts
@@ -1,5 +1,5 @@
-import {$} from "../metadata";
-import {CWCO} from "../cwco";
+import {$} from "../../core/metadata";
+import {CWCO} from "../../cwco";
 
 export function defineNodeContextMetadata(node: Node) {
 	if ($.has(node) && $.get(node)?.$context) {
