+++ src/tracker/utils/resolve-executables.spec.ts
