+++ src/tracker/utils/resolve-executables.ts
@@ -1,7 +1,7 @@
 import {evaluateStringInComponentContext} from "./evaluate-string-in-component-context";
-import {CWCO} from "../cwco";
-import {jsonParse} from "./json-parse";
-import {jsonStringify} from "./json-stringify";
+import {CWCO} from "../../cwco";
+import {jsonParse} from "../../utils/json-parse";
+import {jsonStringify} from "../../utils/json-stringify";
 
 export function resolveExecutables(
 	str: string,
